move powereventwatcher
