+++ ee/powereventwatcher/power_event_watcher_other_test.go
