+++ ee/powereventwatcher/power_event_watcher_windows.go
